requirements.txt
# === Enhanced Python Dependencies for VADAR Autonomous Trading Bot ===

# Core Dependencies
python-dotenv>=1.0.0              # Load environment variables from .env file
requests>=2.31.0                  # HTTP requests for API calls

# LangChain Framework
langchain>=0.1.17                 # LLM agent framework
langchain-groq>=0.1.4            # Groq LLM integration (LLaMA3)

# Data Processing & Analysis
pandas>=2.0.0                     # Data manipulation and analysis
numpy>=1.24.0                     # Numerical computing
tools>=0.1.0                       # Custom tools for data processing

# Logging & Monitoring
rich>=13.7.0                      # Enhanced console output and logging
colorama>=0.4.6                   # Cross-platform colored terminal text

# Web & API
httpx>=0.24.0                     # Modern HTTP client (alternative to requests)
aiohttp>=3.8.0                    # Async HTTP client for concurrent requests

# Data Validation & Types
pydantic>=2.0.0                   # Data validation using Python type annotations
typing-extensions>=4.7.0          # Extended type hints

# Configuration & CLI
click>=8.1.0                      # Command line interface creation
python-decouple>=3.8              # Better environment variable handling

# Optional: Advanced Features
schedule>=1.2.0                   # Job scheduling (alternative to time.sleep loops)
psutil>=5.9.0                     # System and process utilities
tabulate>=0.9.0                   # Pretty-print tabular data

# Optional: Database (if you want to store trading history)
sqlite3                           # Built-in Python SQLite (no pip install needed)
# sqlalchemy>=2.0.0              # SQL toolkit and ORM (uncomment if needed)

# Optional: Visualization (uncomment if you want charts)
# matplotlib>=3.7.0              # Plotting library
# plotly>=5.15.0                 # Interactive plots
# dash>=2.14.0                   # Web dashboard framework

# Optional: Advanced Analytics (uncomment if needed)
# scikit-learn>=1.3.0            # Machine learning library
# ta>=0.10.2                     # Technical analysis library

# Development & Testing (optional)
pytest>=7.4.0                    # Testing framework
black>=23.7.0                    # Code formatter
isort>=5.12.0                    # Import sorter

# === Version Compatibility ===
# Compatible with Python 3.9, 3.10, 3.11, and 3.12
# Groq integration requires API key from https://console.groq.com
# Recall Network API requires registration at https://recall.network

# === Installation Notes ===
# 1. Create virtual environment: python -m venv trading_bot_env
# 2. Activate: source trading_bot_env/bin/activate (Linux/Mac) or trading_bot_env\Scripts\activate (Windows)
# 3. Install: pip install -r requirements.txt
# 4. Copy .env.example to .env and configure your API ke